Adjust CI checks
