+++ jest-ci.config.cjs
@@ -1,0 +1,10 @@
+// Copyright 2017-2022 @polkadot/apps authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+const config = require('@polkadot/dev/config/jest.cjs');
+
+module.exports = {
+  ...config,
+  moduleNameMapper: {},
+  testTimeout: 2 * 60 * 1000
+};
